Refactor components to use centralized export (at client/components/index.js)
