+++ client/components/Landing.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Splash from './Splash';
+import { Splash } from '../components';
 import { Header } from 'semantic-ui-react';
 
 const Landing = (props) => {
