+++ client/components/index.js
@@ -10,5 +10,6 @@ export { default as Footer } from './Footer';
 export { default as Albums } from './Albums';
 export { default as Artists } from './Artists';
 export { default as Landing } from './Landing';
+export { default as Splash } from './Splash';
 export { default as Songs } from './Songs';
 export { Login, Signup } from './AuthForm';
